third_party/blink/web_tests/http/tests/images/document-policy-lossy-images-max-bpp-expected.txt

CONSOLE ERROR: Document policy violation: lossy-images-max-bpp is not allowed in this document.
